uim-rfkill: Change the permission for /dev/tifm
Since android doesn't have a FM user as such, and the current android
fm GUI application comes up with a APP_* UID, modify the fm device's
permissions to 0666.
TODO: figure out the right way to change the owner of the device to proper
UID when it is assigned to the FM UI application.

Change-Id: Idab5192818551c3888968fb839b72e2d45585272
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>