hardware:ti:omap3: uim: remove ppoll
The ppoll used in uim was based on compat_ppoll
implementation from the bluez stack.
The compat_ppoll implementation is nothing but poll itself
which waits indefinitely and not truly ppoll.

So move to use poll and drop the inclusion of a bluez header
un-necessarily.
Also remove the now unwanted signal set.

Change-Id: I949909177488c27ad1932f54267156dd307ff74d
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>