wpan: fix FM Tx get operations
FM character device driver converts the incoming CH8 data from the chip
into HCI-Event data as expected from the FM stack (MCPv2.5).
During this convertion, the parameter length field or rlen is not made
use of in HCI-Event packet and is always valued at 0.

This patch on receiving such data for the frequency commands does the
necessary convertion by multiplying with 10, since the input frequency
to chip during set frequency is divided by a factor 10.
Also correct the comment about the extra zero, it is rather the un-used
rlen field.

resolves DRT: OMAPS00235954

Change-Id: I5ecee352165c119d00d1874e88fe91a73b985c56
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>